refactor(apply): fold.rs を役割別サブモジュールへ分割する
merge=deep 追加で肥大化した fold.rs を、locals/ や hooks/ と同じ「薄い domain-root +
役割別サブモジュール」の形へ分割する。fold.rs は mod 宣言のみの入口とし、
fold/text.rs（concat）・fold/json.rs（shallow/deep）・fold/plist.rs（shallow/deep）へ
ロジックとテストを移す。モジュール境界に形式名が付いたため、関数名の重複プレフィックス
（json_shallow 等）を削り fold::json::shallow のように呼べるようにする。
